fix(node): wrong const type

+++ opentron/src/constants.rs
@@ -21,7 +21,7 @@ pub const MAX_NUM_OF_STANDBY_WITNESSES: usize = 127;
 // 27 * 70% = 18.9, so a solid block is one verified by 19 witnesses.
 pub const SOLID_THRESHOLD_PERCENT: usize = 70;
 
-pub const NUM_OF_SKIPPED_SLOTS_IN_MAINTENANCE: usize = 2;
+pub const NUM_OF_SKIPPED_SLOTS_IN_MAINTENANCE: i64 = 2;
 
 /// Renamed: WitnessAllowanceFrozenTime
 pub const NUM_OF_FRONZEN_DAYS_FOR_WITNESS_ALLOWANCE: usize = 1;
